tests/wpt/web-platform-tests/offscreen-canvas/fill-and-stroke-styles/2d.gradient.object.compare.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by tools/gentest.py. -->
<title>OffscreenCanvas test: 2d.gradient.object.compare</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/canvas-tests.js"></script>

<h1>2d.gradient.object.compare</h1>
<p class="desc"></p>


<script>
var t = async_test("");
t.step(function() {

var offscreenCanvas = new OffscreenCanvas(100, 50);
var ctx = offscreenCanvas.getContext('2d');

var g1 = ctx.createLinearGradient(0, 0, 100, 0);
var g2 = ctx.createLinearGradient(0, 0, 100, 0);
_assertDifferent(g1, g2, "g1", "g2");
ctx.fillStyle = g1;
_assertSame(ctx.fillStyle, g1, "ctx.fillStyle", "g1");

t.done();

});
</script>
